[llvm-exegesis] Fix compilation on lld-x86_64-darwin13
YAMLTraits does not know how to serialize `size_t` portably. Use `int`
instead.

llvm-svn: 329176

+++ llvm/tools/llvm-exegesis/lib/BenchmarkResult.h
@@ -38,7 +38,7 @@ struct InstructionBenchmark {
   AsmTemplate AsmTmpl;
   std::string CpuName;
   std::string LLVMTriple;
-  size_t NumRepetitions = 0;
+  int NumRepetitions = 0;
   std::vector<BenchmarkMeasure> Measurements;
   std::string Error;
